float to double

+++ Client/Shared/Shared.Models/StockPredictionResult.cs
@@ -8,18 +8,18 @@ namespace CapitalClue.Frontend.Shared.Models;
 
 public class StockPredictionResult
 {
-    public float Gain { get; set; }
-    public float ConfidenceLowerGain { get; set; }
-    public float ConfidenceUpperGain { get; set; }
+    public double Gain { get; set; }
+    public double ConfidenceLowerGain { get; set; }
+    public double ConfidenceUpperGain { get; set; }
 
-    public Dictionary<int, float> ForeCastIndex { get; set; }
-    public Dictionary<int, float> ConfidenceLowerBound { get; set; }
-    public Dictionary<int, float> ConfidenceUpperBound { get; set; }
+    public Dictionary<int, double> ForeCastIndex { get; set; }
+    public Dictionary<int, double> ConfidenceLowerBound { get; set; }
+    public Dictionary<int, double> ConfidenceUpperBound { get; set; }
 
     public StockPredictionResult()
     {
-        ForeCastIndex = new Dictionary<int, float>();
-        ConfidenceLowerBound = new Dictionary<int, float>();
-        ConfidenceUpperBound = new Dictionary<int, float>();
+        ForeCastIndex = new Dictionary<int, double>();
+        ConfidenceLowerBound = new Dictionary<int, double>();
+        ConfidenceUpperBound = new Dictionary<int, double>();
     }
 }
